DX-1071: Move CommandLineArgumentError into the errors directory
