+++ .docker/entrypoint/lib/errors/command_line_argument_error.rb
