+++ .docker/entrypoint/spec/includes.rb
@@ -12,6 +12,7 @@ require 'docker_image'
 require 'entrypoint'
 require 'errors/directory_not_found_error'
 require 'errors/file_not_found_error'
+require 'errors/command_line_argument_error'
 require 'exec_env'
 require 'fileutils'
 require 'gemfile_differ'
